// plasma-ui/build/openapi/model/dataSourceDTO.ts
/**
 * PLASMA Data Source Service Application
 * Service for managaing data sources
 *
 * The version of the OpenAPI document: 0.0.3
 * Contact: pomp@uni-wuppertal.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DataSourceDTO {
    uuid?: string;
    title?: string;
    description?: string;
    longDescription?: string;
}